+++ Template/config/duplicatemod_config.php
@@ -19,20 +19,17 @@
 
     </div>
 
-    <!-- <br>
+    <br>
     <br>
 
     <div class="task-form-container">
 
         <div class="task-form-main-column">
-            <?= $this->form->label(t('Sticky time box at the bottom'), 'time_box_enabled') ?>
-            <?= $this->form->checkbox('time_box_enabled', t('enabled'), 1, $time_box_enabled, '', [
-                'autofocus',
-                'tabindex="3"'
-            ]) ?>
+            <?= $this->form->label(t('[Need WeekHelper plugin!!] Add one week to the week pattern in the tasktitle.'), 'weekadd_enabled') ?>
+            <?= $this->form->checkbox('weekadd_enabled', t('enabled'), 1, $weekadd_enabled) ?>
         </div>
 
-    </div> -->
+    </div>
 
 
     <div class="task-form-bottom">